Move HSC repo files to repo/ for better separation
I've separated the CFHT gen3 repo files, the exports.yaml, and the refcats,
so that the repo/ path can be safely entirely deleted when reprocessing the
data. This gets us part way to that layout, and we can move the refcats out
when we re-process the HSC data to get the new table schema.